Add --no-shell arg for non-interactive nixpkgs-review
Usually we want to interactively review, but in some workflows we may want to
use `nixpkgs-review` to batch process a number of PR reviews.

This commit adds a `--no-shell` flag which, when passed, will do everything
except drop into the interactive `nix-shell`, and instead just finish with an
exit code signaling success or failure.