Refactor the report.markdown() function
This commit has no functional changes; it merely refactors the
`Report.markdown()` function to be a pure function returning a `str`, so that we
can thread it through to a `request` post in a future update as part of #89.

Along the way, this adds test coverage to the function to help catch refactoring
bugs, issues like #88, and make the final report output format easier to
view (e.g., can just open the test file to see what it looks like when reviewing
changes to the format).