Do not fetch tags when fetching PR branches
The command to fetch the PR branches is also fetching all reachable tags
from those references, but it's not something that it should be expected
to do. In one-shot fetchs like that the tags are not relevant, as
discussed in the man page of `git tag`.

Add the `--no-tags` flag to disable automatic tag following.

Closes #222.