change behavior to use `nom` by default
Now, nixpkgs-review will check for `nom` in `PATH` and
use it by default. If it isn't found in `$PATH` it will
default to `nix`.

This can be overridden by the `nom-path` flag, which will take
either the path to `nom` (or a similar build utility) or an empty
string.
If an empty string is passed, it will default to `nix`

Signed-off-by: Florian Brandes <florian.brandes@posteo.de>