Update instructions for GitHub API tokens
In late 2022, GitHub [introduced][1] a new type of API tokens that are supposed
to be more secure and fine grained than the old type of tokens.  But they won't
work for nixpkgs-review's purposes (at least for now) because there is no way to
grant these new tokens access to resources that are not owned by you, and most
users of nixpkgs-reviews are not owners of @NixOS/nixpkgs which is the repo
they're trying to submit pull request comments to (when using `--post-result`,
anyway).

This commit clarifies the README instructions so that users will know they need
to create the old type of API token, the "personal access token (classic)", and
not the new type.

Also did a bit of rewording for further clarity.

[1]: https://github.blog/2022-10-18-introducing-fine-grained-personal-access-tokens-for-github/