fix: distinguish build failures from store failures
"nix-store --verify-path" exits with status 1 in various situations,
including when the path does not exist (due to having failed to build)
or when the checksum doesn't match the Nix database.

Since nixpkgs-review tries to report build failures specifically, use
"nix store verify" instead with its "--no-contents" flag. This avoids
classifying checksum mismatches as build failures, which can be
misleading to nixpkgs-review users.

Though still an abnormal situation that might be useful to report, we
now ignore store path errors; these are generally an abnormal
situation, and detecting them is out-of-scope for nixpkgs-review.  Should
it become clearer that it would be useful to report them, it could be
done by removing --no-contents and checking "nix store verify"'s exit
status, which indicates the nature of the problem.