Support --include-pr to merge additional PR into reviewed worktree
Sometimes a PR can be fully validated only with another unmerged PR. In
this scenario, one either has to wait for the first PR to merge, or
construct an alternative test path that is not using nixpkgs-review.

This patch solves it by adding support for --include-pr option that, if
used, will merge additional PRs into the worktree under test.

The final report will list the new option to avoid confusion or
misleading the reader as to the final state under test.

The option can be passed multiple times to include multiple PRs. The
order of merges of the requested PRs is retained.