Combine head_commit/staged data into one variant
Clarify that "the changes to be tested" can EITHER take the form of a
commit OR can be uncommitted changes by using one "changes" variable
where applicable instead of a pair of variables where exactly one is not
None.

Also introduce an enum to name the different forms of uncommitted
changes instead of using a bool.