Isolate git operations with a new repository
Shallow fetch operations permanently change the shallow configuration of
the repository they run in. Using a temporary worktree does not protect
the user's repository against such modifications. Switch to using a
new separate repository to provide appropriate protection.

Set up the alternates mechanism to have the new repository share storage
with the original. This cannot be done automatically by git-clone since
git-clone refuses to use a shallow repository as an alternate object
store (for good reason!). Manually create the new repository and provide
it with a copy of the original's shallow configuration to make this use
of alternates correct.